+++ tests/integrationtests/conftest.py
@@ -1,8 +1,0 @@
-import pytest
-
-from tests.integrationtests.utils import get_geojson_fixture
-
-
-@pytest.fixture
-def europe():
-    return get_geojson_fixture("europe.geojson")