+++ tests/unittests/api/test_response_models.py
@@ -7,7 +7,6 @@ from ohsome_quality_api.api.response_models import (
     IndicatorMetadata,
     IndicatorMetadataCoverageResponse,
     IndicatorMetadataResponse,
-    ProjectMetadataResponse,
     QualityDimensionMetadataResponse,
     TopicMetadata,
     TopicMetadataResponse,
@@ -25,11 +24,9 @@ def test_topic_metadata():
     TopicMetadata(
         name="foo",
         description="foo",
-        endpoint="elements",
         aggregation_type="area",
         filter="foo",
         indicators=["mapping-saturation"],
-        projects=["core"],
         source=None,
     )
 
@@ -40,11 +37,9 @@ def test_topic_metadata_response():
             "building-count": TopicMetadata(
                 name="foo",
                 description="foo",
-                endpoint="elements",
                 aggregation_type="area",
                 filter="foo",
                 indicators=["mapping-saturation"],
-                projects=["core"],
                 source=None,
             )
         }
@@ -89,34 +84,10 @@ def test_metadata_quality_dimensions_list(quality_dimensions):
     assert response.result == quality_dimensions
 
 
-def test_metadata_projects(metadata_project_core):
-    response = ProjectMetadataResponse(result=metadata_project_core)
-    assert response.result == metadata_project_core
-
-
-def test_metadata_projects_fail(project_core):
-    with pytest.raises(ValidationError):
-        ProjectMetadataResponse(result="")
-    with pytest.raises(ValidationError):
-        ProjectMetadataResponse(result="bar")
-    with pytest.raises(ValidationError):
-        ProjectMetadataResponse(result={})
-    with pytest.raises(ValidationError):
-        ProjectMetadataResponse(result={"foo": "bar"})
-    with pytest.raises(ValidationError):
-        ProjectMetadataResponse(result={"foo": project_core})
-
-
-def test_metadata_projects_list(projects):
-    response = ProjectMetadataResponse(result=projects)
-    assert response.result == projects
-
-
 def test_indicator_metadata():
     IndicatorMetadata(
         name="foo",
         description="foo",
-        projects=["core"],
         quality_dimension="completeness",
     )
 
@@ -127,14 +98,13 @@ def test_indicator_metadata_response():
             "mapping-saturation": IndicatorMetadata(
                 name="foo",
                 description="foo",
-                projects=["core"],
                 quality_dimension="completeness",
             )
         }
     )
 
 
-def test_indicator_metadata_response_fail(metadata_indicator_minimal):
+def test_indicator_metadata_response_fail(metadata_indicator_mapping_saturation):
     with pytest.raises(ValidationError):
         IndicatorMetadataResponse(result="")
     with pytest.raises(ValidationError):
@@ -144,7 +114,7 @@ def test_indicator_metadata_response_fail(metadata_indicator_minimal):
     with pytest.raises(ValidationError):
         IndicatorMetadataResponse(result={"foo": "bar"})
     with pytest.raises(ValidationError):
-        IndicatorMetadataResponse(result={"foo": metadata_indicator_minimal})
+        IndicatorMetadataResponse(result={"foo": metadata_indicator_mapping_saturation})
 
 
 def test_indicator_metadata_coverage(bpolys):